Renamed restart.py appropriately
