+++ kano_updater/ui/views/finish.py
